docs: 2주차 미션 기능명세서 작성
