+++ src/main/java/org/weekly/weekly/ui/writer/SystemWriter.java
@@ -1,0 +1,52 @@
+package org.weekly.weekly.ui.writer;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.stereotype.Component;
+import org.weekly.weekly.voucher.domain.DiscountType;
+import org.weekly.weekly.util.PrintMessage;
+import org.weekly.weekly.util.VoucherMenu;
+
+import java.util.Arrays;
+
+@Component
+@ConditionalOnProperty(value="command.read", havingValue = "system")
+public class SystemWriter implements CommandWriter {
+    private final Logger logger = LoggerFactory.getLogger(SystemWriter.class);
+    private void println(String msg) {
+        System.out.println(msg);
+    }
+    private void print(String msg) {System.out.print(msg);}
+
+    public void printVoucherProgram() {
+        println(PrintMessage.EMPTY.getMessage());
+        println(PrintMessage.PROGRAM.getMessage());
+        Arrays.stream(VoucherMenu.values())
+                .forEach(voucherMenu -> System.out.println(voucherMenu.getPrintMsg()));
+    }
+
+    public void printErrorMsg(String errorMsg) {
+        logger.warn(errorMsg);
+        println(PrintMessage.EMPTY.getMessage());
+        println(errorMsg);
+    }
+
+    public void printCreateVoucher(DiscountType discountType) {
+        println(PrintMessage.EMPTY.getMessage());
+        println(PrintMessage.CREATE_VOUCHER.getMessage() + discountType.getInputExampleMessage());
+        print(PrintMessage.INPUT_MESSAGE.getMessage());
+    }
+
+    public void printSelectDiscount() {
+        println(PrintMessage.EMPTY.getMessage());
+        println(PrintMessage.DISCOUNT_SELECT.getMessage());
+        Arrays.stream(DiscountType.values())
+                .forEach(discountMap -> System.out.println(discountMap.getSelectMessage()));
+    }
+
+    public void printReuslt(String result) {
+        println(PrintMessage.EMPTY.getMessage());
+        println(result);
+    }
+}
